[0527]Frontend - React, React router, Zustand, Chakra UI

+++ backend/controllers/product.controller.js
@@ -48,12 +48,16 @@ const updateProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id" })
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" })
     } catch (err) {
         console.log("Error in deleting product:", err.message);
-        res.status(400).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server error" });
     }
 }
 module.exports = { getProducts, createProduct, updateProduct, deleteProduct }
